fix #46 - download deezer favs only downloads the first 25 songs
This is a ugly fix. I don't have a long fav playlist to test it. I don't
know what's the max limit value the backend accepts. Let's see if there
are some bug reports in the future. There is a link to the next x songs
in the json. We could iterate over the json response.